// src/utils/enums/BunkerEnum.ts
export enum BunkerQueryKey {
    BUNKERING_PERFORMANCE = "bunkering_performance",
    BUNKER_HANDLING_BARGE_PERFORMANCE = "bunker_handling_barge_performance",
    BUNKER_SUPPLY = "bunker_supply",
    BUNKER_SUPPLY_PERFORMANCE = "bunker_supply_performance",
    BUNKER_SUPPLY_DAILY = "bunker_supply_daily",
    BUNKER_SUPPLY_FORECAST = "bunker_supply_forecast",
    BUNKER_SUPPLY_PORT = "bunker_supply_port",
    BUNKER_SUPPLY_PORT_FORECAST = "bunker_supply_port_forecast",
    BUNKER_HANDLING_TRANSPORTER_RANK = "bunker_handling_transporter_rank",
    BUNKER_HANDLING_VESSEL_RANK = "bunker_handling_vessel_rank",
    BUNKER_HANDLING_LOSS_HISTORY = "bunker_handling_loss_history",
    BUNKER_PLAN_LIST = "bunker_plan_list",
    BUNKER_PLAN_VERIFIED_LIST = "bunker_plan_verified_list",
    BUNKER_PLAN_LIST_EXPORT = "bunker_plan_list_export",
    BUNKER_PLAN_LIST_VERIFIED_EXPORT = "bunker_plan_list_verified_export",
    BUNKER_PLAN_SCHEDULE_EDIT = "bunker_plan_schedule_edit",
    BUNKER_PLAN_SCHEDULE_BATCH_EDIT = "bunker_plan_schedule_bacth_edit",
    UPLOAD_EVIDENCES_BATCH = "upload_evidences_batch",
    BUNKER_PLAN_STATUS_EDIT = "bunker_plan_schedule_status_edit",
    BUNKER_PLAN_DETAIL = "bunker_plan_schedule_detail",
    BUNKER_PLAN_TELEX_INFO = "bunker_plan_telex_info",
}